requirements/base.txt
django>=1.5
djangorestframework==2.3.8
mstranslator
